requirements.txt
customtkinter
sqlmodel
pillow
PyYAML
